Squash: SystemUI: Tile: Lockscreen fixed for 7.0 by @Millosr
SystemUI: Tile: Lockscreen: Fix lockscreen bypass
Disable the Lockscreen tile when we are in the lockscreen and
when this one is securized.

That avoid to bypass the lockscreen (security issue)

This replace state.visible unsupported since 7.0 and
avoid any confusion for the end user because the tile is
still visible but disabled.

SystemUI: improve lockscreen tile behavior by romanbb
* Stay disabled when disabled
* No need to use Keyguard locks, use internal update monitor

Change-Id: Ib1067e6eb2195db2a19c7de8dd9b9bff1a4c4d71

Conflicts:
	packages/SystemUI/res/values/config.xml
	packages/SystemUI/res/values/rr_strings.xml

Conflicts:
	packages/SystemUI/res/values/config.xml
	packages/SystemUI/res/values/rr_strings.xml
	packages/SystemUI/src/com/android/systemui/keyguard/KeyguardViewMediator.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QSTileHost.java